`MessageBusListenerActor` runs the source and receives the result asynchronously **after** the sink is executed.
Since we were waiting for a promise that was completed on sink
materialization, in can happen that the promise is completed before
the monitor is updated.

This commit changes the test to keep trying until the monitor gets
updated, as there no reliable way to wait for the actor to processed
the result from the stream unless we change the actor the send a
message after updating the monitor.

Signed-off-by: Simão Mata <simao.mata@here.com>