Make the Flyway schema table configurable.
Up until version 5.0.0 Flyway used to store the migration historical
info in a table called `schema_version`. From version 5 the table is
called `flyway_schema_history`. Version 6 drops the fallback mechanism
that allowed to still user the old name for the table. We want to migrate
from 4 to 6 but don't want to go through the hassle of renaming the old
schema table.

Related:
  - https://github.com/flyway/flyway/issues/2010
  - https://flywaydb.org/documentation/releaseNotes#5.0.0

Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>