Fix bug when failure that causes subscriber to die
even when `monitorSafe` is used.

This is a rather subtle and hard to track bug.

`monitor.onProcessed` returns a future and `monitorSafe` recovers from
that future. The problem is `monitor.onProcessed` can fail **before**
actually creating the future, so `monitorOnProcessed` will fail before
it returns a future that `monitorSafe` can recover from, so this was
not safe at all.

Changed it to wrap `onProcessed` into a try to really make sure the
listener does not die if something goes wrong with `monitor`.

Signed-off-by: Simão Mata <simao.mata@here.com>