Improve method naming.
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>

Expose error as result of the http call in `ServiceHttpClient`

Before this change, we were allowing the caller of `ServiceHttpClient`
to specify a error handling partial function. This was meant to give
the user more control over the result of the http request when the
request was not successful (not 2xx). However, this turned out to be a
bit of a leaky abstraction, as `ServiceHttpClient` would call that
error handler PF to get a `Future` and return that, but if that PF
returned a successful `Future` and the caller then used `map` on
that `Future`, then the caller would be operating on the result of the
HTTP call twice. This works, but it's not intuitive and could lead to
some bugs and some complex code like:

```
    execHttpWithNamespace2[Unit](namespace, req) {
      case err if err.status == StatusCodes.NotFound => FastFuture.successful(false)
    }.map {
      case resp if resp.httpResponse.status == StatusCodes.NotFound => false
      case _ => true
    }
```

Here `FastFuture.successful(false)` is returned from the error handler, but
that value still returned into the `map` function, as the future is
still successful.

This is not the only instance of this type of problems. This MR
changes the Api of `ServiceHttpClient` to completely expose a
`RemoteServiceError` rather than an entry point for a PF that handles
it, so the caller of `execHttp*` needs to explicitly handle
`Either[RemoteServiceError, R]` to get a `R` and decide what to do
with a possible `Left[RemoteServiceError]`.

Two convenience methods are provided with `import ServiceHttpClient._`
to handle the most common cases of:

1. Getting `R` out of `Future[Either[RemoteServiceError, R]]`. Calling
`.ok` on `Future[Either[RemoteServiceError, R]]` will return a
successful future of `R` if the value is a `Right`, and a failed
future containing `RemoteServiceError` if the value is `Left`.

2. Handling different types of errors in
`Future[Either[RemoteServiceError, R]]`, using `handleErrors`. This is
a common case and is very similar to what was provided previously
through a partial function defined by the caller of the api. In theory
this method is not needed as the caller could simply use methods in
`Future` to unpack the `Either`, but then the caller would always need
to handle `Right` and all the `Left`, which the caller might not care
about, because the methods provided by `Future` all require functions
and not partial functions.

The names for methods in `ServiceHttpClient` were also changed to be
more explicit. `execHttp` is now named `execHttpUnmarshalled` as we
always try to unmarshall the response using the decoder provided
implicitly.

`execHttpFull` returns both the unmarshalled response and the full
`HttpResponse` obtained by the akka http client.

Signed-off-by: Simão Mata <simao.mata@here.com>