Rename some `EcuReplaced` message members.
I'm doing this for convenience. This message will end up in device-registry
and being exposed interleaved with `DeviceUpdateEvent`s. So naming the
class members as in `DeviceUpdateEvent` just makes the whole serialisation
and processing simpler. And since the raison d'être of this event is to be
exposed like that, why not just go ahead and make my life easier.

Relates-to: OTA-5130
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>