Rework the MockMessageBus to keep multiple events per key.
While testing the ECU replacement events, I noticed that the `MockMessageBus`
only keeps the latest event, because when a new event with the same key
arrives, it replaces the previous event. So for example, if a device replaces
two ECUs, I was only able to find the latest ECU replacement event in `received`.
So I had to rework it a bit to fit my tests.

While doing that, I noticed that we're actually always searching by values and
ignoring the keys, so this could be simplified and modeled with a humble `Seq`
or alike.

Also notice this doesn't break the previous `findReceived` API, so all current
usages should still work when bumping libats. This is tested in an upcoming MR
on director that makes use of this new version of the bus.

Relates-to: OTA-5130
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>